refactor: 클래스명 변경 및 추상화 제거
혼동될 수 있는 클래스명을 변경합니다.
복잡성을 줄이기 위해 하나의 구현체만 사용될 가능성이 높은 추상화(인터페이스)를 제거합니다.
